Replace include_once with include
Causing tests to fail in Laravel applications when `ray.php` is used.

In Laravel tests, when Ray is re-instantiated it attempts to load the config file using `include_once`. When calling `include_once` the first time, it will work as expected. On subsequent calls within the same process it will simply return `true` to indicate that the file has already been included. https://www.php.net/manual/en/function.include-once.php

This is a problem in unit tests, as the Laravel container is destroyed and recreated for each test. The first test will run, but subsequent tests will fail as the config is set to `true` instead of the intended array.