Refer to the Docker specific implementation
Add a link to our Docker specific documentation on the Laravel specific page. This makes it clear that more specific setup is needed in order to get Docker worker with Ray when using Laravel.

Also add a small note about why the configuration file should be placed in the root of the repo and when it is OK to place the file in the config/ directory.